Refactor lecture room integration and cleanup
- Remove old lecture room page: deleted app/dashboard/(course)/lecture-room/[course_id]/lecture-room.tsx
- Move and update lecture room logic to app/dashboard/lectures/[courseId]/lecture-room.tsx
- Refactor lectures page to use new LectureRoom component and hooks
- Remove obsolete components/lecture-room.tsx
- Add modular lecture room components: ChatPanel, ControlPanel, DeleteLectureDialog, DetectionAnalysis, LectureRoom, PresentationPanel, VideoFeed
- Add supporting hooks: useCameraStream, useChatManager, useFaceDetection, useLectureManagement, usePresentationManager
- Add types/lecture-room.ts for type safety
- Update .gitignore for new structure and cleanup
